+++ devbie-core/src/test/java/underdogs/devbie/TestApplication.java
@@ -1,0 +1,26 @@
+package underdogs.devbie;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+@SpringBootApplication
+class TestApplication {
+
+    @TestConfiguration
+    public static class TestConfig {
+
+        @PersistenceContext
+        private EntityManager entityManager;
+
+        @Bean
+        public JPAQueryFactory jpaQueryFactory() {
+            return new JPAQueryFactory(entityManager);
+        }
+    }
+}
