--- devbie-core/src/test/java/underdogs/devbie/TestConfig.java
@@ -1,21 +1,0 @@
-package underdogs.devbie;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-@TestConfiguration
-public class TestConfig {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
