+++ devbie-core/src/test/java/underdogs/devbie/recommendation/domain/AnswerRecommendationRepositoryImplTest.java
@@ -3,35 +3,31 @@ package underdogs.devbie.recommendation.domain;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import javax.persistence.EntityManager;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
-@SpringBootTest
+@DataJpaTest
 class AnswerRecommendationRepositoryImplTest {
-
-    @Autowired
-    private EntityManager em;
 
     @Qualifier("answerRecommendationRepository")
     @Autowired
     private RecommendationRepository repository;
 
+    @Autowired
     private JPAQueryFactory jpaQueryFactory;
+
     private AnswerRecommendationRepositoryImpl repositoryImpl;
 
     private AnswerRecommendation recommendation;
 
     @BeforeEach
     void setUp() {
-        jpaQueryFactory = new JPAQueryFactory(em);
         repositoryImpl = new AnswerRecommendationRepositoryImpl(jpaQueryFactory);
 
         recommendation = AnswerRecommendation.of(1L, 1L, RecommendationType.RECOMMENDED);
